+++ vfprint/src/main/java/com/example/vfprint/dto/response/CalculateResponse.java
@@ -4,18 +4,19 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class CalculateResponse {
-    private double price;
+    private BigDecimal price;
     private int quantityPaper;
     private int productSheet;
     private String paperSize;
-    private double processingCost;
-    private double discount;
-    private double paperCost;
-    private double cost;
+    private BigDecimal processingCost;
+    private BigDecimal discount;
+    private BigDecimal paperCost;
+    private BigDecimal cost;
 }
